+++ vue.config.js
@@ -1,4 +1,9 @@
 module.exports = {
+  configureWebpack: {
+    externals: {
+      moment: 'moment'
+    }
+  },
   pwa: {
     workboxOptions: {
       skipWaiting: true